websocket-docker/requirements.txt
websockets==11.0.3
kafka-python==2.0.2